Kept only ci_sanity tests for securitygroup tests
